requirements.txt
certifi==2019.3.9
chardet==3.0.4
Click==7.0
cyrusbus==0.1.0
Flask==1.0.2
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.1
octopus-http==0.6.4
redis==3.2.1
requests==2.21.0
retools==0.4.1
setproctitle==1.1.10
six==1.12.0
tornado==6.0.2
urllib3==1.24.1
Werkzeug==0.15.1
gunicorn

